+++ service/compiled_benchmark_test.go
@@ -141,10 +141,10 @@ func runFindSimilar(b *testing.B, rows int, cols int) {
 	}
 	for i := 0; i < rows; i++ {
 		record := pb.Record{
-			Data: make([]float64, cols),
+			Data: make([]float32, cols),
 		}
 		for j := 0; j < cols; j++ {
-			record.Data[j] = rand.Float64()
+			record.Data[j] = rand.Float32()
 		}
 
 		if err := records.Create(&record); err != nil {
